fix: update current_epoch type from i64 to i32 for consistency with schema

+++ server/src/handlers/mls_chat/bootstrap_reset_group.rs
@@ -132,7 +132,9 @@ pub async fn handle(
     // FOR UPDATE locks the row so concurrent bootstrap calls serialize on it.
     // Returns (group_info, current_epoch) — None if no row matches; if row
     // matches but group_info IS NOT NULL, the bootstrap already happened.
-    let target_row: Option<(Option<Vec<u8>>, i64)> = sqlx::query_as(
+    // current_epoch is INT4 in the schema; decode as i32 (every other reader
+    // — db.rs, models.rs, federation, mls_auth, actors — uses i32 too).
+    let target_row: Option<(Option<Vec<u8>>, i32)> = sqlx::query_as(
         "SELECT group_info, current_epoch FROM conversations \
          WHERE id = $1 AND group_id = $2 \
          FOR UPDATE",
